+++ app/Models/Country.php
@@ -21,6 +21,8 @@ use Illuminate\Database\Eloquent\Model;
  * @method static \Illuminate\Database\Eloquent\Builder|Country whereUpdatedAt($value)
  * @property string $code
  * @method static \Illuminate\Database\Eloquent\Builder|Country whereCode($value)
+ * @property string $currency
+ * @method static \Illuminate\Database\Eloquent\Builder|Country whereCurrency($value)
  * @mixin \Eloquent
  */
 class Country extends Model
@@ -54,6 +56,28 @@ class Country extends Model
     public function setCode(string $code): self
     {
         $this->code = $code;
+        return $this;
+    }
+
+    public function getCurrency(): string
+    {
+        return $this->currency;
+    }
+
+    public function setCurrency(string $currency): self
+    {
+        $this->currency = $currency;
+        return $this;
+    }
+
+    public function getLocale(): string
+    {
+        return $this->locale;
+    }
+
+    public function setLocale(string $locale): self
+    {
+        $this->locale = $locale;
         return $this;
     }
 }