+++ app/Source/Ride/Enum/RideBaseFiltersEnum.php
@@ -10,5 +10,7 @@ enum RideBaseFiltersEnum: string
 
     case FROM_PLACE = 'from_place_id';
     case TO_PLACE = 'to_place_id';
-    case TIME = 'time';
+    case MIN_TIME = 'min_time';
+    case MAX_TIME = 'max_time';
+    case IS_ACCEPTING_PACKAGE = 'is_accepting_package';
 }