+++ app/Source/Ride/Infra/CreateRide/Services/CreateRideService.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace App\Source\Ride\Infra\CreateRide\Services;
 
+use App\Models\Country;
 use App\Models\Ride;
+use App\Source\Localization\Domain\Facade\LocalizationFacade;
+use App\Source\Localization\Infra\Helpers\LocalizationHelper;
 use Illuminate\Support\Carbon;
 
 class CreateRideService
@@ -16,12 +19,14 @@ class CreateRideService
         Carbon $time,
         int $numberOfSeats,
         int $price,
-        ?string $description
+        ?string $description,
+        bool $isAcceptingPackage,
+        Country $country
     ): void {
         $ride = new Ride();
         $ride
             ->setDriverId($driverId)
-            ->setCurrency('EUR')
+            ->setCurrency($this->getCurrency($country))
             ->setPrice($price)
             ->setDescription($description)
             ->setRideTime($time)
@@ -29,11 +34,21 @@ class CreateRideService
             ->setNumberOfSeats($numberOfSeats)
             ->setFromPlaceId($fromPlaceId)
             ->setToPlaceId($toPlaceId)
+            ->setIsAcceptingPackage($isAcceptingPackage)
             ->save();
 
         $userProfile = $ride->driver->profile;
         $userProfile
             ->increaseTotalRidesCount()
             ->save();
+    }
+
+    private function getCurrency(Country $country): string
+    {
+        if ($currency = LocalizationHelper::getCurrency()) {
+            return $currency;
+        }
+
+        return $country->getCurrency();
     }
 }