+++ database/factories/RideFactory.php
@@ -20,7 +20,8 @@ class RideFactory extends Factory
             'description' => $this->faker->sentence,
             'price' => $this->faker->numberBetween(1, 100),
             'number_of_seats' => $this->faker->numberBetween(1, 3),
-            'currency' => 'EUR'
+            'currency' => 'EUR',
+            'is_accepting_package' => $this->faker->boolean()
         ];
     }
 }