+++ database/migrations/2023_04_23_101531_add_currency_and_locale_to_countries_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('countries', function (Blueprint $table) {
+            $table->string('currency', 3)
+                ->after('code')
+                ->default('');
+
+            $table->string('locale', 2)
+                ->after('currency')
+                ->default('');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::table('countries', function (Blueprint $table) {
+            $table->dropColumn(['currency', 'locale']);
+        });
+    }
+};