+++ database/migrations/2023_04_23_102759_add_country_id_to_rides_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use App\Models\Country;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('rides', function (Blueprint $table) {
+            $table->foreignId('country_id')
+                ->after('driver_id')
+                ->default(Country::where('name', 'Croatia')->first()->getId())
+                ->constrained('countries');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::table('rides', function (Blueprint $table) {
+            $table->dropColumn('country_id');
+        });
+    }
+};