+++ database/seeders/RideSeeder.php
@@ -18,14 +18,16 @@ class RideSeeder extends Seeder
     {
         foreach (User::get() as $user) {
             for ($i = 0; $i < 10; $i++) {
-                $time = ($i === 0) ? Carbon::now()->addDays(rand(1, 10)) : Carbon::now()->subDays(rand(1, 10));
+                $now = Carbon::now()->setTimezone('Europe/Zagreb');
+                $time = ($i === 0) ? $now->addDays(rand(1, 10)) : $now->subDays(rand(1, 10));
 
                 Ride::factory()->state(
                     [
                         'driver_id' => $user->getId(),
                         'from_place_id' => Place::where('name', 'Zagreb')->first()->getId(),
                         'to_place_id' => Place::where('name', 'Split')->first()->getId(),
-                        'time' => $time->addMinutes(rand(1, 10))
+                        'time' => $time->addMinutes(rand(1, 10)),
+                        'time_utc' => $now->utc(),
                     ]
                 )->create();