+++ resources/views/components/checkmark.blade.php
@@ -1,0 +1,5 @@
+@if($type === 'tick')
+    <i class="fa-solid fa-circle-check" style="color: #008f02;"></i>
+@else
+    <i class="fa-sharp fa-solid fa-circle-xmark" style="color: #bd0000;"></i>
+@endif