+++ resources/views/ride/create/create-form.blade.php
@@ -67,8 +67,9 @@
 
                         <div class="mb-6">
                             <label for="price"
-                                   class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">{{__('Price')}}
-                                (€)</label>
+                                   class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                                {{__('Price')}} {{get_user_currency()}}
+                            </label>
                             <input type="number" id="price" name="price"
                                    value="{{old('price')}}"
                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
@@ -82,6 +83,13 @@
                             <textarea type="number" id="description" name="description"
                                       class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
                             >{{old('description')}}</textarea>
+                        </div>
+
+                        <div class="mb-6">
+                            <label for="is_accepting_package"
+                                   class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">{{__('Do you accept packages?')}}
+                            </label>
+                            <x-input type="checkbox" id="is_accepting_package" name="is_accepting_package"/>
                         </div>
 
                         <div class=" mb-6">