+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.24.0] - 2023-04-23
+
+### Changed
+
+- Change country instead of language. Save more country data to the session
+
+## [1.23.0] - 2023-04-23
+
+### Added
+
+- Currency to the country table
+- Country id to the rides table
+
+## [1.22.0] - 2023-04-22
+
+### Added
+
+- "Is accepting packages" feature
+- Fill en.json with all strings from hr.json
+
 ## [1.21.0] - 2023-04-22
 
 ### Added